Revert "code cleanup in models"

+++ app/models/concerns/camaleon_cms/categories_tags_for_posts.rb
@@ -20,8 +20,8 @@ module CamaleonCms::CategoriesTagsForPosts extend ActiveSupport::Concern
   end
 
   # update category assignations for this post
-  # remove assignations that no longer exist
-  # add new assignations
+    # remove assignations that no longer exist
+    # add new assignations
   # cats: (Array) array of category ids assigned for this post, sample: [1,2,3]
   def update_categories(cats=[])
     rescue_extra_data
@@ -29,32 +29,28 @@ module CamaleonCms::CategoriesTagsForPosts extend ActiveSupport::Concern
     old_categories = categories.pluck("#{CamaleonCms::TermTaxonomy.table_name}.id")
     delete_categories = old_categories - cats
     news_categories =  cats - old_categories
-    if  delete_categories.present?
-      term_relationships.where("term_taxonomy_id in (?)", delete_categories).destroy_all
-    end
+    term_relationships.where("term_taxonomy_id in (?)", delete_categories ).destroy_all   if  delete_categories.present?
     news_categories.each do |key|
-      term_relationships.create(term_taxonomy_id: key)
+      term_relationships.create(:term_taxonomy_id => key)
     end
     update_counters("categories")
   end
 
   # update tags assignations for this post
-  # remove assignations that no longer exist
-  # add new assignations
+    # remove assignations that no longer exist
+    # add new assignations
   # tags: (String) tags name separated by commas, sample: "Tag1,Tag two,tag new"
   def update_tags(tags)
     rescue_extra_data
     tags = tags.split(",").strip
-    post_tags = post_type.post_tags
+    post_tags = self.post_type.post_tags
     post_tags = post_tags.where("name not in (?)", tags) if tags.present?
-    term_relationship
-      .where("term_taxonomy_id in (?)", post_tags.pluck("#{CamaleonCms::TermTaxonomy.table_name}.id"))
-      .destroy_all
+    self.term_relationships.where("term_taxonomy_id in (?)", post_tags.pluck("#{CamaleonCms::TermTaxonomy.table_name}.id")).destroy_all
     tags.each do |f|
-      post_tag = post_type.post_tags.where({name: f}).first_or_create(slug: f.parameterize)
-      term_relationships.where({ term_taxonomy_id: post_tag.id }).first_or_create
+      post_tag = self.post_type.post_tags.where({name: f}).first_or_create(slug: f.parameterize)
+      self.term_relationships.where({term_taxonomy_id: post_tag.id}).first_or_create
     end
-    update_counters('tags')
+    update_counters("tags")
   end
 
   # Assign this post for category with id: category_id
@@ -63,9 +59,9 @@ module CamaleonCms::CategoriesTagsForPosts extend ActiveSupport::Concern
     categories_id = [categories_id] if categories_id.is_a?(Integer)
     rescue_extra_data
     categories_id.each do |key|
-      term_relationships.where(term_taxonomy_id: key).first_or_create!
+      term_relationships.where(:term_taxonomy_id => key).first_or_create!
     end
-    update_counters('categories')
+    update_counters("categories")
   end
 
   # Assign this post for category with id: category_id
@@ -73,8 +69,8 @@ module CamaleonCms::CategoriesTagsForPosts extend ActiveSupport::Concern
   def unassign_category(categories_id)
     categories_id = [categories_id] if categories_id.is_a?(Integer)
     rescue_extra_data
-    term_relationships.where(term_taxonomy_id: categories_id).destroy_all
-    update_counters('categories')
+    term_relationships.where(:term_taxonomy_id => categories_id).destroy_all
+    update_counters("categories")
   end
 
   # Assign new tags to this post
@@ -83,10 +79,10 @@ module CamaleonCms::CategoriesTagsForPosts extend ActiveSupport::Concern
     update_counters_before
     tags = tag_titles.split(",").strip
     tags.each do |t|
-      post_tag = post_type.post_tags.where(name: t).first_or_create!
-      term_relationships.where({ term_taxonomy_id: post_tag.id }).first_or_create!
+      post_tag = self.post_type.post_tags.where(name: t).first_or_create!
+      self.term_relationships.where({term_taxonomy_id: post_tag.id}).first_or_create!
     end
-    update_counters('tags')
+    update_counters("tags")
   end
 
   # Unassign tags from this post
@@ -94,8 +90,7 @@ module CamaleonCms::CategoriesTagsForPosts extend ActiveSupport::Concern
   def unassign_tags(tag_titles)
     update_counters_before
     tags = tag_titles.split(",").strip
-    term_relationships
-	  .where({term_taxonomy_id: post_type.post_tags.where(name: tags).pluck(:id)}).destroy_all
+    self.term_relationships.where({term_taxonomy_id: self.post_type.post_tags.where(name: tags).pluck(:id)}).destroy_all
     update_counters("tags")
   end
 
@@ -106,38 +101,32 @@ module CamaleonCms::CategoriesTagsForPosts extend ActiveSupport::Concern
   end
 
   private
-
   @cats_before, @tags_before = [], []
   # save as a cache previous categories and tags assigned for this post
   def rescue_extra_data
-    @cats_before = categories.pluck(:id) if manage_categories?
-    @tags_before = post_tags.pluck(:id) if manage_tags?
+    @cats_before = self.categories.pluck(:id) if manage_categories?
+    @tags_before = self.post_tags.pluck(:id) if manage_tags?
   end
 
   # update quantity of posts assigned for tags and categories assigned to this post
   # if kind is empty, then this will update both: cats and tags
   # kind: (string) tags | categories
   def update_counters(kind = "")
-    if ["", "categories"].include?(kind) && manage_categories?
-      post_type.full_categories.where(id: (@cats_before + categories.pluck(:id)).uniq)
-        .each { |c| c.update_column("count", c.posts.published.size) }
-    end
-    if ["", "tags"].include?(kind) && manage_tags?
-      post_type.post_tags.where(id: (@tags_before + post_tags.pluck(:id)).uniq)
-        .each { |tag| tag.update_column('count', tag.posts.published.size) }
-    end
+    self.post_type.full_categories.where(id: (@cats_before + self.categories.pluck(:id)).uniq).each { |c| c.update_column("count", c.posts.published.size) } if ["", "categories"].include?(kind) && manage_categories?
+    self.post_type.post_tags.where(id: (@tags_before + self.post_tags.pluck(:id)).uniq).each { |tag| tag.update_column("count", tag.posts.published.size) } if ["", "tags"].include?(kind) && manage_tags?
   end
 
   # save extra data such as: categories and tags assigned to this post
   def save_extra_data
-    update_categories(data_categories) if manage_categories? && !data_categories.nil?
-    update_tags(data_tags) if manage_tags? && !data_tags.nil?
+    update_categories(self.data_categories) if manage_categories? && !self.data_categories.nil?
+    update_tags(self.data_tags) if manage_tags? && !self.data_tags.nil?
   end
 
   # auto assign default categories if none categories were assigned
   def check_default_category
     if manage_categories?
-      assign_category([post_type.default_category.id]) unless categories.present?
+      assign_category([self.post_type.default_category.id]) unless self.categories.present?
     end
   end
+
 end
